requirements.txt
fastapi~=0.110.0
requests~=2.31.0
PyYAML~=6.0.1
schedule~=1.2.1